+++ app/controllers/tuttle/execjs_controller.rb
@@ -1,0 +1,11 @@
+require_dependency 'tuttle/application_controller'
+require 'execjs/version' if defined?(::ExecJS)
+
+module Tuttle
+  class ExecjsController < ApplicationController
+
+    def index
+    end
+
+  end
+end