+++ test/controllers/tuttle/execjs_controller_test.rb
@@ -1,0 +1,12 @@
+require 'test_helper'
+
+module Tuttle
+  class ExecjsControllerTest < ActionController::TestCase
+
+    test 'should get index' do
+      get :index
+      assert_response :success
+    end
+
+  end
+end